+++ contrib/fmm_tomography/waveTracker.py
@@ -6,7 +6,7 @@ import numpy as np
 import matplotlib.pyplot as plt
 from PIL import Image
 
-from cofi_espresso.utils import absolute_path as path, silent_remove
+from espresso.utils import absolute_path as path, silent_remove
 
 #--------------------------------------------------------------------------------------------
 
@@ -462,7 +462,7 @@ def run_fm2dss(wdir):
     return subprocess.run(command,stdout=subprocess.PIPE, text=True,shell=True,cwd=wdir)
 
 def compile_fm2dss():
-    # https://github.com/inlab-geo/espresso/blob/main/tools/build_package/validate.py#L170
+    # https://github.com/inlab-geo/espresso/blob/main/espresso_machine/build_package/validate.py#L170
     build_dir = path(".")
     res1 = subprocess.call(["cmake", "."], cwd=build_dir)
     if res1: